Create mk_main_sequence.m
generate a 'main-sequence' type relation between two parameters.
Method - using mk_meanMedian_bin to bin the data in the x-parameter and find the mean and stddev  of the y-parameter in the bin.
Then begin an iterative process of removing outliers in each bin - points beyond a factor fac of the standard deviation within the bin.
The iteration stops when no more points are removed (can be set)